reafactor(#295): CQRS 패턴 적용
